BigDecimal migration + Categories test

+++ src/main/java/fr/client/cacount/Cacount.java
@@ -1,11 +1,13 @@
 package fr.client.cacount;
 
+import java.math.BigDecimal;
+
 /**
  * Created by svirch_n on 12/02/17.
  */
 public class Cacount {
 
     public static final String[] CATEGORY_LIST = {"Alimentaire", "Transport", "Logistique", "Soirée", "Autre", "Soin/Santé"};
-    public static double RATIO = 10.21;
+    public static BigDecimal RATIO = BigDecimal.valueOf(10.21);
 
 }
